+++ CHANGELOG.md
@@ -1,3 +1,53 @@
 ## 0.0.1
 
 - Initial version: setting up skeleton
+
+## 0.0.2
+
+- Fixed http package version
+
+## 0.0.3
+
+- Minor improvements to stability
+
+## 0.0.4
+
+- Implemented ChatAction.fromString
+
+## 0.0.5
+
+- Improvements to ChatMember object
+
+## 0.0.6
+
+- Improvements to Telegram API Client
+
+## 0.0.7
+
+- Minor fixes on start() function
+
+# 0.0.8
+
+- Fixed event loop, now it's not using sleep anymore
+
+# 0.0.9
+
+- Improved bot loop, now it's not blocking and it's a correct Future
+
+# 0.0.9+1
+
+- Small improvement for previous version
+
+# 0.0.9+2
+
+- Added the possibility to chose if the http client should be only closed or restarted
+
+# 0.0.9+3
+
+- Bug fix for event loop
+
+# 0.1.0
+
+- Added functions to download files
+- Fixed File size from String to int
+- Added small test